move swing timer to the ui part, leaving the core of jbotsim independant
