+++ src/jbotsim/Clock.java
@@ -14,9 +14,6 @@ package jbotsim;
 import jbotsim.event.ClockListener;
 import jbotsim.ui.JTopology;
 
-import javax.swing.*;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.HashMap;
@@ -26,30 +23,32 @@ class Clock {
     Topology tp;
     HashMap<ClockListener, Integer> listeners=new HashMap<ClockListener, Integer>();
     HashMap<ClockListener, Integer> countdown=new HashMap<ClockListener, Integer>();
-    Timer timer=new Timer(10, new ActionHandler());
     Integer time=0;
 
     Clock(Topology topology){
         this.tp = topology;
-        timer.start();
+
         try {Thread.sleep(100);} catch (InterruptedException e) {e.printStackTrace();}
     }
-    private class ActionHandler implements ActionListener{
-        public void actionPerformed(ActionEvent evt) {
-            // Delivers messages first
-            tp.getMessageEngine().onClock();
-            // Then give the hand to the nodes
-            tp.getNodeScheduler().onClock(tp);
-            // Then to the topology itself
-            tp.onClock();
-            // Finally, to all other listeners whose countdown has expired
-            for (ClockListener cl : getExpiredListeners()) {
-                cl.onClock();
-                countdown.put(cl, listeners.get(cl)); // reset countdown
-            }
-            time++;
-        }
+
+    /**
+     * Performs a single round
+     */
+    public void step() {
+      // Delivers messages first
+      tp.getMessageEngine().onClock();
+      // Then give the hand to the nodes
+      tp.getNodeScheduler().onClock(tp);
+      // Then to the topology itself
+      tp.onClock();
+      // Finally, to all other listeners whose countdown has expired
+      for (ClockListener cl : getExpiredListeners()) {
+          cl.onClock();
+          countdown.put(cl, listeners.get(cl)); // reset countdown
+      }
+      time++;
     }
+
     protected ArrayList<ClockListener> getExpiredListeners(){
         ArrayList<ClockListener> expiredListeners = new ArrayList<ClockListener>();
         for(ClockListener cl : listeners.keySet()) {
@@ -80,27 +79,15 @@ class Clock {
         countdown.put(listener, 1);
     }
     /**
-     * Unregisters the specified listener. (The <tt>onClock()</tt> method of this 
-     * listener will not longer be called.) 
+     * Unregisters the specified listener. (The <tt>onClock()</tt> method of this
+     * listener will not longer be called.)
      * @param listener The listener to unregister.
      */
     public void removeClockListener(ClockListener listener){
         listeners.remove(listener);
         countdown.remove(listener);
     }
-    /**
-     * Returns the time unit of the clock, in milliseconds.
-     */
-    public int getTimeUnit(){
-        return timer.getDelay();
-    }
-    /**
-     * Sets the time unit of the clock to the specified value in millisecond.
-     * @param delay The desired time unit (1 corresponds to the fastest rate)
-     */
-    public void setTimeUnit(int delay){
-        timer.setDelay(delay);
-    }
+
     /**
      * Returns the current time of the clock in time units.
      */
@@ -108,26 +95,6 @@ class Clock {
         return time;
     }
     /**
-     * Indicates whether the clock is currently running or paused.
-     * @return <tt>true</tt> if running, <tt>false</tt> if paused.
-     */
-    public boolean isRunning(){
-        return timer.isRunning();
-    }
-    /**
-     * Pauses the clock (freezes time and stops to send onClock() events to 
-     * listeners).
-     */
-    public void pause(){
-        timer.stop();
-    }
-    /**
-     * Resumes the clock if it was paused. 
-     */
-    public void resume(){
-        timer.start();
-    }
-    /** 
      * Sets the clock time to 0.
      */
     public void reset(){
